wip: change api and deprecate old way
In the future this will have settings for the ui
